ECS/Visualizer/Spiral
I did significant reworking of the ECS to take advantage of X-Macros,
generating a significant amount of boilerplate. I also added another
style to my visualizer scene, and am working towards generating OpenGL
boilerplate, again using macros. Finally, I am moving my "spiral"
experiment into "galaxy volume", and integrating it into the space
scene. Eventually it will be a regular component in the ECS.

+++ configuration/lua_configuration.h
@@ -1,6 +1,7 @@
 #ifndef LUA_CONFIGURATION_LUA_H
 #define LUA_CONFIGURATION_LUA_H
-#include <lua.h>
+#include <lua-5.3.5/src/lua.h>
+#include <lua-5.3.5/src/lauxlib.h>
 #include <stdbool.h>
 
 //Runs the Lua file located at filepath, and prints errors as appropriate.
@@ -42,7 +43,7 @@ char * getoptfieldstr(lua_State *L, int i, const char *var, const char *d);
 //Generic get field value, based on the type of the default value.
 #define getoptfield(L, index, var, d) _Generic((d), \
 	bool:         getoptfieldbool, \
-	int:          (int)getoptfieldint, \
+	int:          getoptfieldint, \
 	lua_Integer:  getoptfieldint, \
 	lua_Number:   getoptfieldnum, \
 	float:        getoptfieldnum, \
